Testing1/requirements.txt
numpy==2.3.1
pandas==2.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
openai==1.12.0
google-generativeai==0.8.3
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
PyPDF2==3.0.1
